backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9

# Validation
pydantic==2.5.2
pydantic-settings==2.1.0

# AI APIs
anthropic==0.7.7
google-generativeai==0.3.1

# HTTP Client
httpx==0.25.2

# Authentication (Phase 2)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Integrations
twilio==8.11.0
sendgrid==6.11.0

# Task Queue (Optional - Phase 2)
redis==5.0.1
celery==5.3.4

# Utilities
python-dotenv==1.0.0
